% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMatch.R
\name{allMatch}
\alias{allMatch}
\title{Match all elements in vector.}
\usage{
allMatch(query, subject)
}
\description{
Return all matching indexes in subject vector from query vector.
}
\examples{
# Input query and subject vecots
match.list <- allMatch(query = c(1,2,3), subject = c(1,1,2,2,3,3,2))
# Output is query length list with subject indexes
match.list
# $`1`
# [1] 1 2
#
# $`2`
# [1] 3 4 7
#
# $`3`
# [1] 5 6
}

