% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPackages.R
\name{cleanPackages}
\alias{cleanPackages}
\title{Remove all non-default packages from session.}
\usage{
cleanPackages(verbose = TRUE)
}
\arguments{
\item{verbose}{print out summary, default is \code{TRUE}}
}
\description{
Remove all non-default packages from session.
}
\author{
Daniel Leongamornlert
}
\keyword{clean}
\keyword{detach}
\keyword{package}
\keyword{session}

