% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRsid.R
\name{extractRsid}
\alias{extractRsid}
\title{Extract rsid}
\usage{
extractRsid(x, sepIn = ":", sepOut = ":")
}
\arguments{
\item{x}{character vector of variant IDs}

\item{sepIn}{separator used in varinat IDs, default ":"}

\item{sepOut}{separator used in output varinat IDs, where there is no rsid, default ":"}
}
\value{
a character vector of variant IDs
}
\description{
This function extracts rsid from variant id
}
\examples{
extractRsid(c("1:154:CTT:C", "rs79438587:1543:C:T", "rs7"))
}
\author{
Tokhir Dadaev
}
\keyword{SNP}
\keyword{extract}
\keyword{impute2}
\keyword{rsid}
