% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFieldPtRep.R
\name{FFieldPtRep}
\alias{FFieldPtRep}
\title{Force field simulation for a set of points}
\usage{
FFieldPtRep(coords, rep.fact = 20, rep.dist.lmt = 10, attr.fact = 0.2,
  adj.max = 0.1, adj.lmt = 0.5, iter.max = 10000)
}
\arguments{
\item{coords}{matrix or data.frame consisting of two columns (x and y coordinates).}

\item{rep.fact}{repulsion force factor.}

\item{rep.dist.lmt}{repulsion distance limit.}

\item{attr.fact}{attraction force factor.}

\item{adj.max}{maximum position adjustment at each iteration.}

\item{adj.lmt}{position adjustment limit at which the simulation stops.}

\item{iter.max}{the maximum number of iterations beyond which simulation will end and a warning will be reported.}
}
\description{
Force field simulation of interaction of set of points. Adapted form FField v0.1.0 package by Grigori Kapoustin on 27/04/2006
}
\seealso{
\url{http://cran.r-project.org/web/packages/FField}
}

