% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frr.R
\name{frr}
\alias{frr}
\title{Calculate Familial Relative Risk (FRR)}
\usage{
frr(data, FamilialRisk = 2)
}
\arguments{
\item{data}{is data.frame object with 3 columns, `c("snp", "maf", "est")`.}

\item{FamilialRisk}{Familial risk to first-degree relatives of cancer cases. Default is 2.}
}
\value{
data.frame with FRR and ContributionFRR columns.
}
\description{
This function calculates FRR and Contribution of FRR percentage.
}
\keyword{familial}
\keyword{frr}
\keyword{risk}
\keyword{snp}
