% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen2dose.R
\name{gen2dose}
\alias{gen2dose}
\title{Convert IMPUTE2 gen to dose}
\usage{
gen2dose(genFile, sampleFile, chrName)
}
\arguments{
\item{genFile}{input gen file}

\item{sampleFile}{if provided sampleID will be added as header, else X1..Xn as sampleID}

\item{chrName}{chromosome name \code{c(1:22,X,Y)}}
}
\value{
a \code{data.frame} object
}
\description{
This function converts IMPUTE2 gen file posterior probabilities to dosage: collapse every 3 genetype columns for each sample into one columns as: 0*AA + 1*AB + 2*BB
}
\author{
Tokhir Dadaev
}
\keyword{convert}
\keyword{dosage}
\keyword{dose}
\keyword{gen}
\keyword{impute2}
