% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSymbol.R
\name{geneSymbol}
\alias{geneSymbol}
\title{Get genes as GRanges object}
\usage{
geneSymbol(chrom = NA, chromStart = NA, chromEnd = NA)
}
\arguments{
\item{chrom}{chromosome name, must be character class with length(chrom)==1, e.g.: chr1"}

\item{chromStart, chromEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}
}
\value{
a GRanges object
}
\description{
This function returns GRanges object, used as input for plotting gene track
Warning: too much dependency to other packages, will need to rewrite/update the code in future far far away...
Warning: only works for LE at the moment with pre-loaded bio packages
}
\author{
Tokhir Dadaev
}
\keyword{gene}
\keyword{granges}
\keyword{plot}
\keyword{symbol}
