% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSummary.R
\name{glmSummary}
\alias{glmSummary}
\title{A formatted dataframe output for glm() object.}
\usage{
glmSummary(model = NA, description = NA, summary = FALSE)
}
\arguments{
\item{model}{Output of glm() Defaults to NA.}

\item{description}{Description column, string, Defaults to NA.}

\item{summary}{condensed summary Defaults to FALSE.}
}
\description{
This function allows you to get pretty GLM
}
\examples{
glmSummary()
}
\keyword{glm}
\keyword{output}
\keyword{summary}

