% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperlink.R
\name{hyperlink}
\alias{hyperlink}
\title{LocusExplorer hyperlinks for SNPs and geneSymbols}
\usage{
hyperlink(id = NULL, type = "SNP", link = NULL, html = TRUE)
}
\arguments{
\item{id}{character vector of SNP rsids or geneSymbols, e.g.: "rs34948850" or "BOK"}

\item{type}{link type, "SNP" for SNPs and "geneSymbol" for genes. Default "SNP".}

\item{link}{Default, for type is SNP then NCBI link, type is geneSymbol then UCSC, or provide custom link.}

\item{html}{Default is TRUE, wrapped the link as HTML.}
}
\description{
LocusExplorer ggplot custom theme.
}
\author{
Tokhir Dadaev
}
\keyword{NCBI}
\keyword{hyperlink}
\keyword{locusexplorer}
