% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICRcolours.R
\name{ICRcolours}
\alias{ICRcolours}
\title{ICR colours}
\usage{
ICRcolours(col = "Primary", plot = FALSE)
}
\arguments{
\item{col}{numeric value to match rownumber, or Pallete name, or Colour name.}

\item{plot}{logical, defaults to FALSE}
}
\value{
a \code{character vector} object with hex values
}
\description{
Returns ICR brand colour Hex values
}
\author{
Tokhir Dadaev
}
\keyword{colour}
\keyword{hex}
\keyword{icr}
