% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBIrsidHyperlink.R
\name{NCBIrsidHyperlink}
\alias{NCBIrsidHyperlink}
\title{LocusExplorer ggplot custom theme}
\usage{
NCBIrsidHyperlink(SNP,
  link = "https://www.ncbi.nlm.nih.gov/projects/SNP/snp_ref.cgi?rs=",
  html = TRUE)
}
\description{
LocusExplorer ggplot custom theme.
}
\author{
Tokhir Dadaev
}
\keyword{NCBI}
\keyword{hyperlink}
\keyword{locusexplorer}
