% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnnot.R
\name{plotAnnot}
\alias{plotAnnot}
\title{LocusExplorer - Annotation plot, collapsed}
\usage{
plotAnnot(
  data,
  chrom = NULL,
  xStart = NULL,
  xEnd = NULL,
  vline = NULL,
  collapse = FALSE,
  pad = TRUE
)
}
\arguments{
\item{data}{dataframe with c(CHR, BP, TYPE1, TYPE2, COLOUR_HEX, TYPE2N) columns.}

\item{chrom}{chromosome name, interger 1:23}

\item{xStart, xEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}

\item{vline}{Mark hits on genes}

\item{collapse}{Collapse TYPE2 group into TYPE1 group}

\item{pad}{Default is TRUE, to align plots pad strings with spaces, using oncofunco::strPadLeft().}
}
\value{
a \code{ggplot} object
}
\description{
Annotation plot for LocusExplorer.
}
\author{
Tokhir Dadaev
}
\keyword{annotation}
\keyword{locusexplorer}
\keyword{plot}
