% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBlank.R
\name{plotBlank}
\alias{plotBlank}
\title{Blank plot}
\usage{
plotBlank(
  xStart = NULL,
  xEnd = NULL,
  yLabel = "",
  textAnnot = "No data",
  pad = TRUE
)
}
\arguments{
\item{xStart, xEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}

\item{yLabel}{character string, label for Y axis, default "".}

\item{textAnnot}{character string for plot text annotation, default "No data"}

\item{pad}{Default is TRUE, to align plots pad strings with spaces, using oncofunco::strPadLeft().}
}
\value{
a \code{ggplot} object
}
\description{
Blank plot for LocusExplorer, if data is missing blank output ggplot with text annotation.
}
\author{
Tokhir Dadaev
}
