% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEQTL.R
\name{plotEQTL}
\alias{plotEQTL}
\title{LocusExplorer - TCGA eQTL plot}
\usage{
plotEQTL(
  data = NULL,
  xStart = NULL,
  xEnd = NULL,
  hits = NULL,
  hitsTags = TRUE,
  label = "eQTL TCGA",
  pad = TRUE,
  title = NULL
)
}
\arguments{
\item{data}{data.frame with columns: c("SNP", "gene", "xSNP", "xGene")}

\item{xStart, xEnd}{Region range, minimum and maximum position on x axis, advised to keep this less than 5Mb.}

\item{hits}{SNP names to label in the plot. Must be present in data data.frame.}

\item{hitsTags}{Tag SNP names to label in the plot. Must be present in data data.frame.}

\item{pad}{Default is TRUE, to align plots pad strings with spaces, using oncofunco::strPadLeft().}

\item{title}{character string for plot title. Default is NULL, i.e.: no plot title.}
}
\value{
a \code{ggplot} object
}
\description{
TCGA eQTL plot for LocusExplorer.
}
\author{
Tokhir Dadaev
}
\keyword{SNP}
\keyword{TCGA}
\keyword{eQTL}
\keyword{genetics}
\keyword{plot}
