% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistone.R
\name{plotHistone}
\alias{plotHistone}
\title{LocusExplorer - ENCODE histone track plot}
\usage{
plotHistone(
  folder = NULL,
  file = c("wgEncodeBroadHistoneGm12878H3k27acStdSig.bigWig",
    "wgEncodeBroadHistoneH1hescH3k27acStdSig.bigWig",
    "wgEncodeBroadHistoneHsmmH3k27acStdSig.bigWig",
    "wgEncodeBroadHistoneHuvecH3k27acStdSig.bigWig",
    "wgEncodeBroadHistoneK562H3k27acStdSig.bigWig",
    "wgEncodeBroadHistoneNhekH3k27acStdSig.bigWig",
    "wgEncodeBroadHistoneNhlfH3k27acStdSig.bigWig"),
  name = c("GM12878", "H1-hESC", "HSMM", "HUVEC", "K562", "NHEK", "NHLF"),
  colour = c("#FF1F1F", "#FFB41F", "#2ADFAF", "#1FB4FF", "#1F1FFF", "#B41FFF",
    "#FF1FB4"),
  alpha = 0.7,
  chr = NULL,
  xStart = NULL,
  xEnd = NULL,
  pad = TRUE,
  title = NULL
)
}
\arguments{
\item{folder}{folder path for bigWig files}

\item{file}{bigWig file(s) names, defaults to encode filenames.}

\item{name}{defaults to encode names: c("GM12878","H1-hESC","HSMM","HUVEC","K562","NHEK","NHLF")}

\item{colour}{colour for bigWig files, defaults to encode colours: c("#FF1F1F","#FFB41F","#2ADFAF","#1FB4FF","#1F1FFF","#B41FFF","#FF1FB4")}

\item{alpha}{transparency for overlayed tracks, default 0.7.}

\item{chr, xStart, xEnd}{subset bigWig, chromosome number, postion start and end.}

\item{pad}{Default is TRUE, to align plots pad strings with spaces, using oncofunco::strPadLeft().}

\item{title}{character string for plot title. Default is NULL, i.e.: no plot title.}
}
\value{
a \code{ggplot} object
}
\description{
ENCODE Histone plot for LocusExplorer.
H3K27Ac Mark (Often Found Near Active Regulatory Elements) on 7 cell lines from ENCODE 
http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeRegMarkH3k27ac/
}
\author{
Tokhir Dadaev
}
\keyword{ENCODE}
\keyword{bigWig}
\keyword{histone}
\keyword{plot}
