% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLD.R
\name{plotLD}
\alias{plotLD}
\title{LocusExplorer - LD plot}
\usage{
plotLD(data = NULL, xStart = NULL, xEnd = NULL, hits = NULL,
  hitsLabel = TRUE, pad = TRUE, title = NULL)
}
\arguments{
\item{data}{plink LD output format, data.frame object with c("BP_A","SNP_A","BP_B","SNP_B","R2") columns. Optional/recommended.}

\item{xStart, xEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}

\item{hits}{SNP names to label in the plot. Must be present in assoc data.frame.}

\item{hitsLabel}{Default is TRUE, set to FALSE not to show SNP names on the plot.}

\item{pad}{Default is TRUE, to align plots pad strings with spaces, using oncofunco::strPadLeft().}

\item{title}{character string for plot title. Default is NULL, i.e.: no plot title.}
}
\value{
a \code{ggplot} object
}
\description{
LD plot for LocusExplorer.
}
\author{
Tokhir Dadaev
}
\keyword{LD}
\keyword{SNP}
\keyword{genetics}
\keyword{plot}
