% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLDarc.R
\name{plotLDarc}
\alias{plotLDarc}
\title{LocusExplorer - LDarc plot}
\usage{
plotLDarc(
  LD = NULL,
  xStart = NULL,
  xEnd = NULL,
  upper = NULL,
  lower = NULL,
  statNames = c("Stepwise", "JAM"),
  minR2 = 0.4,
  hitsOnly = FALSE,
  pad = TRUE,
  title = NULL
)
}
\arguments{
\item{LD}{plink LD output format, data.frame object with c("BP_A","SNP_A","BP_B","SNP_B","R2") columns.}

\item{xStart, xEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}

\item{upper, lower}{SNP names to label in the plot. Must be present in assoc data.frame.}

\item{statNames}{Two statistics methods name used for y-axis, string, default c("Stepwise", "JAM").}

\item{minR2}{Filter LD on minimum R2, filter out "non-significant" relationships.}

\item{pad}{Default is TRUE, to align plots pad strings with spaces, using oncofunco::strPadLeft().}

\item{title}{character string for plot title. Default is NULL, i.e.: no plot title.}
}
\value{
a \code{ggplot} object
}
\description{
LD Arc plot for LocusExplorer.
}
\author{
Tokhir Dadaev
}
\keyword{LD}
\keyword{SNP}
\keyword{genetics}
\keyword{plot}
