% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLDmatrix.R
\name{plotLDmatrix}
\alias{plotLDmatrix}
\title{LocusExplorer - LD plot matrix heatmap}
\usage{
plotLDmatrix(data = NULL, xStart = NULL, xEnd = NULL, hits = NULL,
  hitsLabel = TRUE, title = NULL)
}
\arguments{
\item{data}{plink LD output format, data.frame object with c("BP_A","SNP_A","BP_B","SNP_B","R2") columns.}

\item{xStart, xEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}

\item{hits}{SNP names to label in the plot. Must be present in assoc data.frame.}

\item{title}{character string for plot title. Default is NULL, i.e.: no plot title.}
}
\value{
a \code{ggplot} object
}
\description{
LD plot for LocusExplorer.
}
\author{
Tokhir Dadaev
}
\keyword{LD}
\keyword{SNP}
\keyword{genetics}
\keyword{plot}
