% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMiami.R
\name{plotMiami}
\alias{plotMiami}
\title{Miami plot}
\usage{
plotMiami(dat = NULL, hits = NULL, hitsName = hits,
  hitsLabel = TRUE, hitsColour = NULL, pMinFilter = 0.05,
  pMaxCap = 1/10^50, yRange = c(-50, 50), yRangeBy = 5,
  colP1 = c("green", "limegreen"), colP2 = c("lightblue", "blue"),
  title = c("My Main Title", "My Top Title", "My Bottom Title"))
}
\arguments{
\item{dat}{SNP association results, data.frame object with c("SNP","BP","P") columns. Required.}

\item{hits}{SNP names to label in the plot. Must be present in assoc data.frame.}

\item{hitsName}{alternative SNP names to label in the plot. Default same as `hits`}

\item{hitsLabel}{Default is TRUE, set to FALSE not to show SNP names on the plot.}

\item{hitsColour}{Default NULL, uses ggplot colours.}

\item{yRange}{}

\item{yRangeBy}{y-axis ticks, setting to 5 means ticks will be placed at `c(0, 5, 10, ...)`.}

\item{title}{character string for plot title. Default is NULL, i.e.: no plot title.}

\item{yMin, yMax}{y-axis range limits.}
}
\value{
a \code{ggplot} object
}
\description{
Miami plot - WORK IN PROGRESS...
}
\author{
Tokhir Dadaev, Edward J. Saunders
}
\keyword{SNP}
\keyword{genetics}
\keyword{manhattan}
\keyword{miami}
\keyword{plot}
