% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNPtype.R
\name{plotSNPtype}
\alias{plotSNPtype}
\title{LocusExplorer - SNP type plot}
\usage{
plotSNPtype(
  assoc = NULL,
  xStart = NULL,
  xEnd = NULL,
  pad = TRUE,
  title = NULL,
  typedColName = "TYPED",
  typedValue = 2,
  colTyped = "#003333",
  colImputed = "#669999"
)
}
\arguments{
\item{assoc}{SNP association results, data.frame object with c("SNP","BP","P") columns. Required.}

\item{xStart, xEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}

\item{pad}{Default is TRUE, to align plots pad strings with spaces, using oncofunco::strPadLeft().}

\item{title}{character string for plot title. Default is NULL, i.e.: no plot title.}

\item{typedColName}{Column name in assoc data.frame to indicate typed/imputed SNPs, default "TYPED".}

\item{typedValue}{Value for typed SNPs, default is 2.}

\item{colTyped}{Colour for typed SNPs, default "#003333".}

\item{colImputed}{Colour for imputed SNPs, default "#669999".}
}
\value{
a \code{ggplot} object
}
\description{
SNP type plot for LocusExplorer.
}
\author{
Tokhir Dadaev
}
\keyword{SNP}
\keyword{genetics}
\keyword{plot}
