% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadeTintColour.R
\name{shadeTintColour}
\alias{shadeTintColour}
\title{Shade and tints for input colour}
\usage{
shadeTintColour(col = "orange", change = 25)
}
\arguments{
\item{col}{a character string with a valid colour name, see ?col2rgb for valid colour names. Defauls to "orange".}

\item{change}{numeric vector, negative values for tints, positive values for shades.}
}
\value{
character vector of colour names.
}
\description{
Shade and tints for input colour
}
\seealso{
\url{http://colorbrewer2.org/}
}
\author{
Tokhir Dadaev
}
