% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_LE.R
\name{theme_LE}
\alias{theme_LE}
\title{LocusExplorer ggplot custom theme}
\usage{
theme_LE()
}
\description{
LocusExplorer ggplot custom theme.
}
\author{
Tokhir Dadaev
}
\keyword{ggplot}
\keyword{locusexplorer}
\keyword{theme}
