% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueBimVariant.R
\name{uniqueBimVariant}
\alias{uniqueBimVariant}
\title{Fix Plink BIM file variant IDs when missing or duplicated}
\usage{
uniqueBimVariant(fileIn = NULL, fileOut = NULL, sep = "_")
}
\arguments{
\item{fileIn}{Plink BIM filename}

\item{fileOut}{Output filename Plink BIM format}

\item{sep}{Separator to use when variant ID is missing, default "_".}
}
\value{
none, outputs BIM file
}
\description{
This function calculates MAF for imputed SNP data in dosage format.
Related Biostars post: \url{https://www.biostars.org/p/315219/}
BIM file description: \url{https://www.cog-genomics.org/plink2/formats#bim}
}
\author{
Tokhir Dadaev
}
\keyword{bim}
\keyword{convert}
\keyword{plink}
\keyword{snp}
\keyword{variant}
