% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourHue.R
\name{colourHue}
\alias{colourHue}
\title{ggplot2 colours}
\usage{
colourHue(n)
}
\arguments{
\item{n}{number of colours to output}
}
\value{
n number of colour names
}
\description{
Generate colours matching ggplot
}
\seealso{
\url{http://stackoverflow.com/questions/8197559}
}
\author{
Tokhir Dadaev
}
