% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourHue.R
\name{colourHue}
\alias{colourHue}
\title{ggplot2 colours}
\usage{
colourHue(n)
}
\arguments{
\item{n}{number of colours to output.}
}
\value{
n number of colour names.
}
\description{
Generate colours matching ggplot. Adapted from JohnColby's solution at StackOverflow.
}
\seealso{
\url{https://stackoverflow.com/a/8197703/680068}
}
\author{
Tokhir Dadaev
}
