% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen2vcf.R
\name{gen2vcf}
\alias{gen2vcf}
\title{Convert IMPUTE2 gen to VCFv4.1 format, input for fastqtl}
\usage{
gen2vcf(genFile, chrName, outVCFFile, subsetSNPs = NULL,
  subsetSamples = NULL)
}
\arguments{
\item{genFile}{input gen file}

\item{chrName}{chromosome name \code{c(1:22,X,Y)}}

\item{outVCFFile}{output file name for VCF}

\item{subsetSNPs}{index of SNPs to subset - rows}

\item{subsetSamples}{index of Samples to subset - coloumns}
}
\value{
outputs VCF file, nothing returned in the workspace
}
\description{
This function converts IMPUTE2 gen file posterior probabilities to dosage, then outputs to VCF file.
}
\keyword{convert}
\keyword{dosage}
\keyword{dose}
\keyword{fastqtl}
\keyword{gen}
\keyword{impute2}
\keyword{vcf}

