% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMAF.R
\name{getMAF}
\alias{getMAF}
\title{Calculate MAF}
\usage{
getMAF(Z = NULL, NoCall = 9)
}
\arguments{
\item{Z}{matrix object, rows are samples, columns are SNPs, values range 0-2.}

\item{NoCall}{missing value for genotype, defaults to 9.}
}
\value{
a \code{matrix} object. First column is MAF (range 0-0.5), second column is 1 if the MAF is flipped, else 0.
}
\description{
This function calculates MAF for imputed SNP data in dosage format.
}
\examples{
# dummy SNP data, 25 samples, 4 SNPs
set.seed(123)
geno <- matrix(sample(c(0, 1, 2), 100, replace = TRUE), ncol = 4)
# calculate MAF, returns 2 column matrix
getMAF(geno)
}
\keyword{dosage}
\keyword{maf}
\keyword{snp}

