% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSummary.R
\name{glmSummary}
\alias{glmSummary}
\title{A formatted output for glm() object}
\usage{
glmSummary(
  model = NA,
  description = NA,
  summary = FALSE,
  decimal = 2,
  scientificP = TRUE
)
}
\arguments{
\item{model}{Output of \code{glm()}, defaults to \code{NA}.}

\item{description}{Description column, string, defaults to \code{NA}.}

\item{summary}{condensed summary, defaults to \code{FALSE}.}

\item{decimal}{rounding decimals, defaul \code{2}.}

\item{scientificP}{Scientific notation for p-value, default \code{TRUE}.}
}
\value{
a \code{data.frame} object
}
\description{
This function allows you to get pretty GLM summary table.
}
\examples{
fit <- glm(mpg ~ cyl + wt, data = mtcars)
glmSummary(Model = fit)
glmSummary(Model = fit, Description = "testing")
glmSummary(Model = fit, Description = "testing", Summary = TRUE)
}
\author{
Tokhir Dadaev
}
\keyword{glm}
\keyword{output}
\keyword{summary}
