% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSummary.R
\name{glmSummary}
\alias{glmSummary}
\title{A formatted output for glm() object}
\usage{
glmSummary(Model = NA, Description = NA, Summary = FALSE)
}
\arguments{
\item{Model}{Output of \code{glm()}, defaults to \code{NA}.}

\item{Description}{Description column, string, defaults to \code{NA}.}

\item{summary}{condensed summary, defaults to \code{FALSE}.}
}
\value{
a \code{data.frame} object
}
\description{
This function allows you to get pretty GLM summary table.
}
\examples{
fit <- glm(mpg ~ cyl + wt, data = mtcars)
glmSummary(Model = fit)
glmSummary(Model = fit, Description = "testing")
glmSummary(Model = fit, Description = "testing", Summary = TRUE)
}
\keyword{glm}
\keyword{output}
\keyword{summary}

