% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGene.R
\name{plotGene}
\alias{plotGene}
\title{Plot genes using ggbio::geom_alignment}
\usage{
plotGene(chrom = NULL, chromStart = NULL, chromEnd = NULL, vline = NULL,
  pad = TRUE)
}
\arguments{
\item{chrom}{chromosome name, must be character class with length(chrom)==1, e.g.: chr1"}

\item{chromStart, chromEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}

\item{vline}{Mark hits on genes}

\item{pad}{Default is TRUE, to align plots pad strings with spaces, using oncofunco::strPadLeft().}
}
\value{
a list with 2 objects: $genePlot \code{ggplot} object and $geneCount numeric
}
\description{
This function returns GRanges object, used as input for plotting gene track
Warning: too much dependency to other packages, will need to rewrite/update the code in future far far away...
Warning: only works for LE at the moment with pre-loaded bio packages
}
\author{
Tokhir Dadaev
}
\keyword{gene}
\keyword{granges}
\keyword{plot}
\keyword{symbol}
