% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLDmatrix.R
\name{plotLDmatrix}
\alias{plotLDmatrix}
\title{LocusExplorer - LD plot matrix heatmap}
\usage{
plotLDmatrix(
  data = NULL,
  xStart = NULL,
  xEnd = NULL,
  hits = NULL,
  hitsName = hits,
  hitsLabel = TRUE,
  hitsNameLength = 20,
  cols = c("lightblue", "#fef0d9", "#fdcc8a", "#fc8d59", "#e34a33", "#b30000"),
  title = NULL,
  legend = TRUE
)
}
\arguments{
\item{data}{plink LD output format, data.frame object with c("BP_A","SNP_A","BP_B","SNP_B","R2") columns.}

\item{xStart, xEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}

\item{hits}{SNP names to label in the plot. Must be present in assoc data.frame.}

\item{hitsName}{Alternative SNP names to label in the plot. Default same as `hits`}

\item{hitsLabel}{Default is TRUE, set to FALSE not to show SNP names on the plot.}

\item{hitsNameLength}{Maximum characters for the label.}

\item{cols}{colours, length 6.}

\item{title}{Character string for plot title. Default is NULL, i.e.: no plot title.}

\item{legend}{Show plot legend, default TRUE.}
}
\value{
a \code{ggplot} object
}
\description{
LD plot for LocusExplorer.
}
\author{
Tokhir Dadaev
}
\keyword{LD}
\keyword{SNP}
\keyword{genetics}
\keyword{plot}
