% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManhattan.R
\name{plotManhattan}
\alias{plotManhattan}
\title{LocusExplorer - Manhattan plot}
\usage{
plotManhattan(assoc = NULL, LD = NULL, geneticMap = NULL,
  suggestiveLine = 5, genomewideLine = 8, xStart = NULL, xEnd = NULL,
  hits = NULL, hitsName = hits, hitsLabel = TRUE, pad = TRUE,
  postprob = FALSE, title = NULL, opts = c("Recombination", "LD",
  "LDSmooth", "SuggestiveLine", "GenomewideLine", "Hits"))
}
\arguments{
\item{assoc}{SNP association results, data.frame object with c("SNP","BP","P") columns. Required.}

\item{LD}{plink LD output format, data.frame object with c("BP_A","SNP_A","BP_B","SNP_B","R2") columns. Optional/recommended.}

\item{geneticMap}{Recombination map, data.frame object with c("BP", "RECOMB") columns. Subset of one of genetic_map_*_combined_b37.txt, at http://mathgen.stats.ox.ac.uk/impute/1000GP_Phase3/ . Optional.}

\item{suggestiveLine}{Suggestive line, default is 5.}

\item{genomewideLine}{Genomewide link, dafault is 8.}

\item{xStart, xEnd}{Region range, zoom, minimum BP and maximum BP, advised to keep this less than 5Mb.}

\item{hits}{SNP names to label in the plot. Must be present in assoc data.frame.}

\item{hitsName}{alternative SNP names to label in the plot. Default same as `hits`}

\item{hitsLabel}{Default is TRUE, set to FALSE not to show SNP names on the plot.}

\item{pad}{Default is TRUE, to align plots pad strings with spaces, using oncofunco::strPadLeft().}

\item{postprob}{Default is FALSE, used for LocusExplorer to plot JAM PostProbs instead of Pvalues.}

\item{title}{character string for plot title. Default is NULL, i.e.: no plot title.}

\item{opts}{Default is c("Recombination","LD","LDSmooth","SuggestiveLine","GenomewideLine","Hits"), parts of plot to display.}
}
\value{
a \code{ggplot} object
}
\description{
Manhattan plot for LocusExplorer.
}
\author{
Tokhir Dadaev
}
\keyword{SNP}
\keyword{genetics}
\keyword{manhattan}
\keyword{plot}
